Passes the Recess shape from the gds file to the mesh flask generator. the recess type from Create your own mesh and the Upload GDS are now passed to the mesh page and are displayed within the side bar on the right. When the user tests their selected mesh the recess shape has been included in the event we want to use COMSOL application builder this can be used to define the type of recess shape.  This commit also has a new change with the upload gds containing new booleans. What I have found in previous testing is that if we want to traverse back and forth, we will need to have the unique dictionary sites saved. Before hand we had 2 boolean checks which were ran_unique_sites which would check to see if there are images within the gallery (captured images are exectued after the identification of unique bondsites and remain in a folder) and the input fields from the user have been filled out. Now we have 2 new booleans we are introducing which are the recess_type and the info_content. Recess type ensures that the user has selected a recess formation from the drop down menu within the side bar, and the info content looks at the html and sees if there is any text in there (if there is text then the unique bondsite information is present).
I will have to think about a way to preserve the dictionary of unique sites as I want to have the unique site preserved in the event the user wants to go back and change parameters.

Another thing that I will have to consider is the issue with the thumbnails caputre. As mentioned yesterday by Dhruva, the way that the thumbnails is captured is by using the KLayout command which uses the personal directory. We need to figure out a way on how to make this more generalized.